requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
et-xmlfile==1.1.0
idna==3.7
lxml==5.2.2
openpyxl==3.1.5
pillow==10.4.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
python-docx==1.1.2
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
